+++ packages/contracts/src/systems/_RegistryCreateRelationshipSystem.sol
@@ -16,12 +16,12 @@ contract _RegistryCreateRelationshipSystem is System {
 
   function execute(bytes memory arguments) public onlyOwner returns (bytes memory) {
     (
-      uint256 npcIndex,
+      uint32 npcIndex,
       uint256 relIndex,
       string memory name,
       uint256[] memory whitelist,
       uint256[] memory blacklist
-    ) = abi.decode(arguments, (uint256, uint256, string, uint256[], uint256[]));
+    ) = abi.decode(arguments, (uint32, uint256, string, uint256[], uint256[]));
     uint256 registryID = LibRegRel.get(components, npcIndex, relIndex);
 
     require(registryID == 0, "Registry: Relationship already exists");
@@ -34,7 +34,7 @@ contract _RegistryCreateRelationshipSystem is System {
   }
 
   function executeTyped(
-    uint256 npcIndex,
+    uint32 npcIndex,
     uint256 relIndex,
     string memory name,
     uint256[] memory whitelist,
